+++ common/src/main/java/net/spell_engine/api/spell/event/SpellEvents.java
@@ -22,7 +22,7 @@ public class SpellEvents {
         @Nullable SpellCast.Attempt onCastingAttempt(Args args);
     }
 
-    public static final Event<SpellCostConsumeEvent> SPELL_COST_CONSUME = new Event<>();
+    public static final Event<SpellCostConsumeEvent> COST_CONSUME = new Event<>();
     public interface SpellCostConsumeEvent {
         record Args(PlayerEntity caster, RegistryEntry<Spell> spell, ItemStack itemStack) {}
         void onSpellCostConsume(Args args);